% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighttpd.R
\name{ovva_install_lighttpd}
\alias{ovva_install_lighttpd}
\title{Install lighttpd}
\usage{
ovva_install_lighttpd(force = FALSE)
}
\arguments{
\item{force}{logical: force reinstallation if lighttpd already exists}
}
\value{
the path to the installed executable
}
\description{
This is a helper function to install lighttpd. Currently it only works on Windows platforms. The lighttpd bundle will be downloaded from \url{http://lighttpd.dtech.hu/} and saved to your user appdata directory.
}
\examples{
\dontrun{
  ovva_install_lighttpd()
}

}
\references{
\url{http://lighttpd.dtech.hu/}
}
