% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{ovva_playlist_handler}
\alias{ovva_playlist_handler}
\title{The default handler for playlists within the shiny app}
\usage{
ovva_playlist_handler()
}
\value{
A tibble, with columns 'skill', 'specific', and 'fun'. Each row defines a playlist entry that corresponds to specific match conditions. This playlist entry is referred to by the name given in \code{specific}, and applies to the skill in column \code{skill}. The corresponding entry in \code{fun} should be a function that takes three arguments (\code{x}, the play-by-play data; \code{team}, the team names to look for, and \code{player}, the player names to look for) and returns the subset of rows of \code{x} that correspond to the playlist conditions.
}
\description{
You can cook up your own version of this function to suit your liking, and pass it as the \code{playlist_handler} parameter to \code{\link{ovva_shiny}}.
}
\seealso{
\code{\link{ovva_shiny}}
}
