% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{ovva_shiny_demo}
\alias{ovva_shiny_demo}
\title{Start the ovva shiny app with a demonstration data set}
\usage{
ovva_shiny_demo()
}
\description{
The demonstration data set is the 2018 women's final (first of 3) from the German Bundesliga: Allianz MTV Stuttgart vs SSC Palmberg Schwerin.
}
\examples{
\dontrun{
  ovva_shiny_demo()
}

}
\seealso{
\code{\link{ovva_shiny}}
}
