% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{ovva_video_server}
\alias{ovva_video_server}
\title{Start a separate server process for serving video}
\usage{
ovva_video_server(method, port)
}
\arguments{
\item{method}{string: either "lighttpd" or "servr". If "lighttpd" but the lighttpd executable cannot be found, "servr" will be used as a fallback}

\item{port}{integer: the port to open the server on. If missing or \code{NULL}, a random port between 8001 and 12000}
}
\value{
A list with components \code{method}, \code{url}, \code{port}, \code{dir} (the source dir from which files will be served), and \code{cleanup_fun} (a function to run when finished with the server, which will stop the server process and clean up its \code{dir})
}
\description{
Note that performance with "servr" may not be great, particularly with each new video, because servr has to read the entire video file before it can serve clips from it.
}
\seealso{
\code{\link{ovva_shiny}}
}
