% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ovva_shiny}
\alias{ovva_shiny}
\title{Launch the Shiny app}
\usage{
ovva_shiny(
  data_path,
  playlist_handler = ovva_playlist_handler(),
  launch_browser = TRUE,
  ...
)
}
\arguments{
\item{data_path}{character: a named character vector of paths to data files. The names will be used as the competition names. So e.g. \code{c("Competition 1" = "/path/to/dvw/files", "Competition 2" = "/path/to/other/dvw/files")}}

\item{playlist_handler}{tibble: a tibble that provides playlist handler capabilities (see \code{\link{ovva_playlist_handler}} for details)}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny]{runApp}}'s \code{launch.browser} parameter)}

\item{...}{: additional parameters passed to the UI and server functions}
}
\description{
Launch the Shiny app
}
